Patient Factory added, Fix issue when save to PatientOrganization

+++ database/migrations/2022_06_19_154014_create_patients_table.php
@@ -43,7 +43,7 @@ return new class extends Migration {
             $table->string('allergies')->nullable();
             $table->integer('height')->nullable();
             $table->integer('weight')->nullable();
-            $table->integer('bmi')->nullable();
+            $table->float('bmi')->nullable();
             $table->timestamps();
         });
     }
